+++ CHANGES.md
@@ -2,8 +2,7 @@
 * none
 
 **Changed:**
-* optimized MacroIdentifierChecker for checking if data is valid
-* preprocessLine is now clean and works 100$ (except for #s thingy)
+* corrected regex for preprocessor replacement.
 
 **Fixed:**
 * none